bugfix : artikel service

+++ src/utils/services/tag.service.js
@@ -7,27 +7,35 @@ module.exports = {
     return tags;
   },
 
-  getTagById: async (id) => {
-    const tag = await prisma.tag.findUnique({ where: { id: parseInt(id, 10) } });
+  getTagByName: async (data) => {
+    const { nama } = data;
+    const tag = await prisma.tag.findFirst({ where: { nama } });
 
     return tag;
   },
 
   createTag: async (data) => {
-    const tag = await prisma.tag.create({
-      data: {
-        nama: data.nama,
-      },
-    });
+    const { nama } = data;
+
+    const findTag = await prisma.tag.findFirst({ where: { nama } });
+
+    if (findTag) return findTag;
+
+    const tag = await prisma.tag.create({ data: { nama } });
 
     return tag;
   },
 
   updateTag: async (id, data) => {
+    const { nama } = data;
+    const findTag = await prisma.tag.findUnique({ where: { id: parseInt(id, 10) } });
+
+    if (!findTag) {
+      return { status: false, message: `Tag with id ${id} not exist!` };
+    }
+
     const tag = await prisma.tag.update({
-      data: {
-        nama: data.nama,
-      },
+      data: { nama },
       where: { id: parseInt(id, 10) },
     });
 
@@ -35,6 +43,12 @@ module.exports = {
   },
 
   deleteTag: async (id) => {
+    const findTag = await prisma.tag.findUnique({ where: { id: parseInt(id, 10) } });
+
+    if (!findTag) {
+      return { status: false, message: `Tag with id ${id} not exist!` };
+    }
+
     const tag = await prisma.tag.delete({ where: { id: parseInt(id, 10) } });
 
     return tag;
